+++ src/cplscheme/CouplingScheme.cpp
@@ -5,9 +5,9 @@ namespace cplscheme {
 
 const double CouplingScheme::UNDEFINED_TIME = -1.0;
 
-const int CouplingScheme::UNDEFINED_TIMESTEPS = -1;
+const int CouplingScheme::UNDEFINED_TIME_WINDOWS = -1;
 
-const double CouplingScheme::UNDEFINED_TIMESTEP_LENGTH = -1.0;
+const double CouplingScheme::UNDEFINED_TIME_WINDOW_SIZE = -1.0;
 
 } // namespace cplscheme
 } // namespace precice